requirements.txt
anthropic==0.25.0
httpx==0.24.1
gspread
google-auth
schedule
requests
fastapi
uvicorn
passlib[bcrypt]
python-jose[cryptography]
asyncpg
python-multipart
